Remove nullity annotations
IntelliJ instruments null checks in when compiling, which is undesirable as it leads to extra useless checks. A micro-optimization for sure, but the original was so performance-mindful that even the result of String#length is cached.
I don't pretend that my own modifications are as performant, but I try my best.